profile read only init
